# api/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": null,
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:44482/api/v1",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a JWT via given credentials.",
                "description": "Login to User",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the token information.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get auth user info",
                "description": "Auth user info",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Return the user information.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error: Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log the user out (Invalidate the token).",
                "description": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Return the status.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh a token",
                "description": "Refresh token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Return the token information.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Auth": {
                "properties": {
                    "accessToken": {
                        "description": "Access Token",
                        "type": "string"
                    },
                    "tokenType": {
                        "description": "Token Type",
                        "type": "string"
                    },
                    "expiresIn": {
                        "description": "Token Expires in",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "Error list",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Status": {
                "properties": {
                    "status": {
                        "description": "Status of Request",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string"
                    },
                    "admin status": {
                        "description": "User admin status",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Authorize to the API, and use returned token for other requests",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}